Move conf/README to conf/README.md and update its contents.
That file now describes what all of the configuration files in the conf
directory are for, and omits the rather dated information about
global.conf (whatever that file may have been ages ago).  It also
includes barebones instructions on how to run webwork2 for development,
and how to deploy webwork2 for production.

Also add NGINX deployment files and instructions.