Fix utf8 characters in warnings and morbo logs to the terminal.
The 'warnings' stash value is not utf8 encoded.  That was removed with
the switch to Mojolicious since there is no need for it.  That means
that it should not be utf8 decoded when it is used.

The same is true of the 'authen_error' stash value.

There was also a problem created by the `use open qw(:encoding(UTF-8) :std);`
call that was in lib/WeBWorK/File/Classlist.pm.  When `use open` is
called in that way it has global scope and affects all standard handles
for the entire application.  Most importantly it was setting the
encoding for STDERR.  Mojolicious writes already encoded logs to STDERR
when morbo is used.  This results in mojibake in the terminal.  Never
call `use open` in that way.  Instead use `use open IO => ':encoding(UTF-8)'`
to apply encoding only in the lexical scope of where it is called.  Even
better, don't call `use open`.  Just apply it as needed when opening a
file.

Many of the binmode calls also need to be reviewed.  This isn't being
done correctly and there are still issues with this.  This is most
important for PG code going forward.