Add pattern validation to the email address field on the feedback page.
Also add some other attributes to the field that are useful for such a
field (placeholder, autocorrect: off, and autocapitalization: off), and
make the field of type "email" instead of "text" to begin with.

It also wouldn't hurt to add server side validation of this fields
contents in case an email saved for the user in the database is not
valid, but that is not done yet here.