Fix setting user overrides for dates on the user detail page for set versions.
This addresses issue #2340.

Currently if an instructor unchecks a checkbox for a user's versioned
set on the user detail page (the page you can reach by clicking on the
"Assigned Sets" column for a user in the Accounts Manager), then when
dates are checked (in the `checkDates` method of `UserDetail.pm`), the
dates from the global set record are used. If those are valid dates
(they usually will be in this case), then those valid dates are returned
with no error.  Thus the database will be updated.  However, the logic
used to determine how the database is updated does not match the logic
used to check the sets, and since the override checkbox is not checked,
the date in the user's versioned set record is set to NULL.  So for
versioned sets this makes it so that if the checkbox is not checked or
the override date is not set, then the global set date is NOT used for a
fallback.  Instead 0 is used for a fallback.  Since that is not a valid
date, this forces an error, and the database is not updated.

This is actually a more general issue with the logic used when the dates
are checked and the logic used when the database is updated not being
the same.  So that is fixed.

Additionally when the dates in the database are updated, there are cases
where the user interface does not correctly reflect the changes made.
One such case is for a regular set if the checkbox for a date is checked
but the override date left empty.  Then the global set date is used
(meaning the user set value for that date is set to NULL in the
database), however the checkbox remains checked.  So the corresponding
parameters are updated so that the changes are reflected when the page
rerenders after saving.

Note that in the case that the dates are not updated in the database due
to an error in the given dates, the user interface does not update to
reflect the current database state.  I don't think the user interface
should be updated in this case, as the instructor might want an
opportunity to correct the error and not have their changes wiped out.

I also noticed that the "User overrides" header was being displayed over
the "Open", "Closes", "Answer" labels in the case that a set is not
assigned to a user (and so there are no override fields).  So that is
not shown in that case anymore.

I also removed the 10 year limit on dates.  I am not sure why we impose
this limit.  There is no reason to do so.  In fact, it is quite useful
in many situations to have a due date that is more than 10 years in the
future.  By the way, I was curious if these dates were affected by the
year 2038 problem.  So after removing this restriction I set a due date
to be in 2050.  This seems to work with no issue. So it seems that 64bit
integers are being used in this case which are not affected by the year
2038 problem.  So we have 292 million years before this will be an issue
again.