More improvements to the User Detail page.
The header on the user override dates now read "User Overrides" unless
it is a test version in which case it is "User's Test Version Dates".

The header on the dates for the global set reads "Test Dates" for tests
and "Assignment Dates" for other assignments.

The placeholder in the user override date inputs now reads "Test
Default" for tests, and "Assignment Default" for other assignments.

The date override checkboxes have been removed.

Remove the "s" from the end of the "Close" date label.  That was
inconsistent with the other date labels.

The class dates for the set are now in readonly inputs instead of just
being text.  This is the same as on the set detail page.  One advantage
of this is that it makes it easier to cut and paste those dates.

There is also a small tweak to the datepicker.js (which is used on this
page as well as on the sets manager page and the set detail page).  When
flatpickr adds the date picker, it hides the input originally on the
page, and adds another input. This moves the id from the original input
onto the added flatpickr input so that labels still work.  The ids
aren't used for anything else after the date picker javascript has found
them.  Also remove the placeholder that flatpickr copies to the new
input since that isn't valid on a hidden input.