Update the docker build to Ubuntu 24.04.
Note that the imagemagick-allow-pdf-read.patch is no longer needed.
Ubuntu 24.04 allows PDF read rights by default.  The reason this
restriction was previously applied was due to a security vulnerability
in ghostscript.  That vulnerability has long been fixed (even for Ubuntu
22.04 although the policy was not updated).  The patch is left in the
repository for Ubuntu 22.04 users.

Also, the Mojo::SQLite version downgrade is not needed anymore with the
version of Mojo::SQLite in Ubuntu 24.04.