Reset the cookie session before Mojolicious saves it.
Currently the session parameters are set at the beginning of each
request.  However, if another request occurs for the same process,
before the first request completes, then the session for the first
request gets saved with the parameters of the second request.  This is
because there only one Mojoicious session setup for the entire app and
the session parameters are global for the process.

To fix this the session parameters need to be set again at the end of
the request just before the session is saved.

This is only an issue if there are multiple clients per worker process.
Of course that is not the case at this point, but hopefully will be some
day.